[Feat] Player Hp Guage, inventory size System [Isuue] not set
플레이어 HP시스템과
인벤토리에서 아이템 무게를 정해서 인벤토리 최대 크기 지정
[issue] 인벤토리 Debug에서는 저장되지만 ui에서 update 안됨

+++ Assets/Scripts/Player/PlayerStatus.cs
@@ -21,7 +21,7 @@ public class PlayerStatus : MonoBehaviour
     public float rotSpeed;
 
     [SerializeField]
-    public float curPower;
+    public int curPower;
 
     public int maxHp;
